+++ src/sha1.js
@@ -1,5 +1,5 @@
 /*
- * js-sha1 v0.1.3
+ * js-sha1 v0.2.0
  * https://github.com/emn178/js-sha1
  *
  * Copyright 2014-2015, emn178@gmail.com
@@ -10,161 +10,219 @@
 ;(function(root, undefined){
   'use strict';
 
+  var NODE_JS = typeof(module) != 'undefined';
+  if(NODE_JS) {
+    root = global;
+  }
+  var TYPED_ARRAY = typeof(Uint8Array) != 'undefined';
   var HEX_CHARS = '0123456789abcdef'.split('');
+  var EXTRA = [-2147483648, 8388608, 32768, 128];
+  var SHIFT = [24, 16, 8, 0];
 
-  var sha1 = function(message, asciiOnly) {
-    var blocks, h0, h1, h2, h3, h4;
-    if(!asciiOnly && /[^\x00-\x7F]/.test(message)) {
-      blocks = getBlocksFromUtf8(message);
-    } else {
-      blocks = getBlocksFromAscii(message);
-    }
+  var blocks = [];
+
+  Array.prototype.__ARRAY__ = true;
+  if(TYPED_ARRAY) {
+    Uint8Array.prototype.__ARRAY__ = true;
+  }
+
+  var sha1 = function(message) {
+    var h0, h1, h2, h3, h4, block = 0, code, end = false, t, f,
+        i, j, index = 0, start = 0, bytes = 0, length = message.length;
+
     h0 = 0x67452301;
     h1 = 0xEFCDAB89;
     h2 = 0x98BADCFE;
     h3 = 0x10325476;
     h4 = 0xC3D2E1F0;
 
-    for(var i = 0, length = blocks.length;i < length;i += 16) {
-      var w = [], j;
-      for(j = 0;j < 16;++j) {
-        w[j] = blocks[i + j];
-      }
-      for(j = 16;j < 80;++j) {
-        var x = w[j - 3] ^ w[j - 8] ^ w[j - 14] ^ w[j - 16];
-        w[j] = leftrotate(x, 1);
-      }
-
-      var a = h0;
-      var b = h1;
-      var c = h2;
-      var d = h3;
-      var e = h4;
-      var f, tmp;
-
-      for(j = 0;j < 20;++j) {
-        f = (b & c) | ((~b) & d);
-        tmp = leftrotate(a, 5) + f + e + 0x5A827999 + w[j];
-        e = d;
-        d = c;
-        c = leftrotate(b, 30);
-        b = a;
-        a = tmp;
-      }
-
-      for(;j < 40;++j) {
-        f = b ^ c ^ d;
-        tmp = leftrotate(a, 5) + f + e + 0x6ED9EBA1 + w[j];
-        e = d;
-        d = c;
-        c = leftrotate(b, 30);
-        b = a;
-        a = tmp;
-      }
-
-      // k = 0x8F1BBCDC;
-      for(;j < 60;++j) {
-        f = (b & c) | (b & d) | (c & d);
-        tmp = leftrotate(a, 5) + f + e + 0x8F1BBCDC + w[j];
-        e = d;
-        d = c;
-        c = leftrotate(b, 30);
-        b = a;
-        a = tmp;
-      }
-
-      for(;j < 80;++j) {
-        f = b ^ c ^ d;
-        tmp = leftrotate(a, 5) + f + e + 0xCA62C1D6 + w[j];
-        e = d;
-        d = c;
-        c = leftrotate(b, 30);
-        b = a;
-        a = tmp;
-      }
-
-      h0 += a;
-      h1 += b;
-      h2 += c;
-      h3 += d;
-      h4 += e;
-    }
-
-    return toHexString(h0) + toHexString(h1)+ toHexString(h2) + toHexString(h3) + toHexString(h4);
-  };
-
-  var leftrotate = function(x, c) {
-    return (x << c) | (x >>> (32 - c));
-  };
-
-  var toHexString = function(num) {
-    var hex = '';
-    for(var i = 0; i < 4; i++) {
-      var offset = 3 - i << 3;
-      hex += HEX_CHARS[(num >> (offset + 4)) & 0x0F] + HEX_CHARS[(num >> offset) & 0x0F];
-    }
-    return hex;
-  };
-
-  var getBytesFromUtf8 = function(str) {
-    var bytes = [], index = 0;
-    for (var i = 0;i < str.length; i++) {
-      var c = str.charCodeAt(i);
-      if (c < 0x80) {
-        bytes[index++] = c;
-      } else if (c < 0x800) {
-        bytes[index++] = 0xc0 | (c >> 6);
-        bytes[index++] = 0x80 | (c & 0x3f);
-      } else if (c < 0xd800 || c >= 0xe000) {
-        bytes[index++] = 0xe0 | (c >> 12);
-        bytes[index++] = 0x80 | ((c >> 6) & 0x3f);
-        bytes[index++] = 0x80 | (c & 0x3f);
+    do {
+      blocks[0] = block;
+      blocks[16] = blocks[1] = blocks[2] = blocks[3] =
+      blocks[4] = blocks[5] = blocks[6] = blocks[7] =
+      blocks[8] = blocks[9] = blocks[10] = blocks[11] =
+      blocks[12] = blocks[13] = blocks[14] = blocks[15] = 0;
+      if(message.__ARRAY__) {
+        for (i = start;index < length && i < 64; ++index) {
+          blocks[i >> 2] |= message[index] << SHIFT[i++ & 3];
+        }
       } else {
-        c = 0x10000 + (((c & 0x3ff) << 10) | (str.charCodeAt(++i) & 0x3ff));
-        bytes[index++] = 0xf0 | (c >> 18);
-        bytes[index++] = 0x80 | ((c >> 12) & 0x3f);
-        bytes[index++] = 0x80 | ((c >> 6) & 0x3f);
-        bytes[index++] = 0x80 | (c & 0x3f);
+        for (i = start;index < length && i < 64; ++index) {
+          code = message.charCodeAt(index);
+          if (code < 0x80) {
+            blocks[i >> 2] |= code << SHIFT[i++ & 3];
+          } else if (code < 0x800) {
+            blocks[i >> 2] |= (0xc0 | (code >> 6)) << SHIFT[i++ & 3];
+            blocks[i >> 2] |= (0x80 | (code & 0x3f)) << SHIFT[i++ & 3];
+          } else if (code < 0xd800 || code >= 0xe000) {
+            blocks[i >> 2] |= (0xe0 | (code >> 12)) << SHIFT[i++ & 3];
+            blocks[i >> 2] |= (0x80 | ((code >> 6) & 0x3f)) << SHIFT[i++ & 3];
+            blocks[i >> 2] |= (0x80 | (code & 0x3f)) << SHIFT[i++ & 3];
+          } else {
+            code = 0x10000 + (((code & 0x3ff) << 10) | (message.charCodeAt(++index) & 0x3ff));
+            blocks[i >> 2] |= (0xf0 | (code >> 18)) << SHIFT[i++ & 3];
+            blocks[i >> 2] |= (0x80 | ((code >> 12) & 0x3f)) << SHIFT[i++ & 3];
+            blocks[i >> 2] |= (0x80 | ((code >> 6) & 0x3f)) << SHIFT[i++ & 3];
+            blocks[i >> 2] |= (0x80 | (code & 0x3f)) << SHIFT[i++ & 3];
+          }
+        }
       }
-    }
-    return bytes;
+      bytes += i - start;
+      start = i - 64;
+      if(index == length) {
+        blocks[i >> 2] |= EXTRA[i & 3];
+        ++index;
+      }
+      block = blocks[16];
+      if(index > length && i < 56) {
+        blocks[15] = bytes << 3;
+        end = true;
+      }
+
+      for(j = 16;j < 80;++j) {
+        t = blocks[j - 3] ^ blocks[j - 8] ^ blocks[j - 14] ^ blocks[j - 16];
+        blocks[j] =  (t << 1) | (t >>> 31);
+      }
+
+      var a = h0, b = h1, c = h2, d = h3, e = h4;
+      for(j = 0;j < 20;j += 5) {
+        f = (b & c) | ((~b) & d);
+        t = (a << 5) | (a >>> 27);
+        e = t + f + e + 1518500249 + blocks[j] << 0;
+        b = (b << 30) | (b >>> 2);
+
+        f = (a & b) | ((~a) & c);
+        t = (e << 5) | (e >>> 27);
+        d = t + f + d + 1518500249 + blocks[j + 1] << 0;
+        a = (a << 30) | (a >>> 2);
+
+        f = (e & a) | ((~e) & b);
+        t = (d << 5) | (d >>> 27);
+        c = t + f + c + 1518500249 + blocks[j + 2] << 0;
+        e = (e << 30) | (e >>> 2);
+
+        f = (d & e) | ((~d) & a);
+        t = (c << 5) | (c >>> 27);
+        b = t + f + b + 1518500249 + blocks[j + 3] << 0;
+        d = (d << 30) | (d >>> 2);
+
+        f = (c & d) | ((~c) & e);
+        t = (b << 5) | (b >>> 27);
+        a = t + f + a + 1518500249 + blocks[j + 4] << 0;
+        c = (c << 30) | (c >>> 2);
+      }
+
+      for(;j < 40;j += 5) {
+        f = b ^ c ^ d;
+        t = (a << 5) | (a >>> 27);
+        e = t + f + e + 1859775393 + blocks[j] << 0;
+        b = (b << 30) | (b >>> 2);
+
+        f = a ^ b ^ c;
+        t = (e << 5) | (e >>> 27);
+        d = t + f + d + 1859775393 + blocks[j + 1] << 0;
+        a = (a << 30) | (a >>> 2);
+
+        f = e ^ a ^ b;
+        t = (d << 5) | (d >>> 27);
+        c = t + f + c + 1859775393 + blocks[j + 2] << 0;
+        e = (e << 30) | (e >>> 2);
+
+        f = d ^ e ^ a;
+        t = (c << 5) | (c >>> 27);
+        b = t + f + b + 1859775393 + blocks[j + 3] << 0;
+        d = (d << 30) | (d >>> 2);
+
+        f = c ^ d ^ e;
+        t = (b << 5) | (b >>> 27);
+        a = t + f + a + 1859775393 + blocks[j + 4] << 0;
+        c = (c << 30) | (c >>> 2);
+      }
+
+      for(;j < 60;j += 5) {
+        f = (b & c) | (b & d) | (c & d);
+        t = (a << 5) | (a >>> 27);
+        e = t + f + e - 1894007588 + blocks[j] << 0;
+        b = (b << 30) | (b >>> 2);
+
+        f = (a & b) | (a & c) | (b & c);
+        t = (e << 5) | (e >>> 27);
+        d = t + f + d - 1894007588 + blocks[j + 1] << 0;
+        a = (a << 30) | (a >>> 2);
+
+        f = (e & a) | (e & b) | (a & b);
+        t = (d << 5) | (d >>> 27);
+        c = t + f + c - 1894007588 + blocks[j + 2] << 0;
+        e = (e << 30) | (e >>> 2);
+
+        f = (d & e) | (d & a) | (e & a);
+        t = (c << 5) | (c >>> 27);
+        b = t + f + b - 1894007588 + blocks[j + 3] << 0;
+        d = (d << 30) | (d >>> 2);
+
+        f = (c & d) | (c & e) | (d & e);
+        t = (b << 5) | (b >>> 27);
+        a = t + f + a - 1894007588 + blocks[j + 4] << 0;
+        c = (c << 30) | (c >>> 2);
+      }
+
+      for(;j < 80;j += 5) {
+        f = b ^ c ^ d;
+        t = (a << 5) | (a >>> 27);
+        e = t + f + e - 899497514 + blocks[j] << 0;
+        b = (b << 30) | (b >>> 2);
+
+        f = a ^ b ^ c;
+        t = (e << 5) | (e >>> 27);
+        d = t + f + d - 899497514 + blocks[j + 1] << 0;
+        a = (a << 30) | (a >>> 2);
+
+        f = e ^ a ^ b;
+        t = (d << 5) | (d >>> 27);
+        c = t + f + c - 899497514 + blocks[j + 2] << 0;
+        e = (e << 30) | (e >>> 2);
+
+        f = d ^ e ^ a;
+        t = (c << 5) | (c >>> 27);
+        b = t + f + b - 899497514 + blocks[j + 3] << 0;
+        d = (d << 30) | (d >>> 2);
+
+        f = c ^ d ^ e;
+        t = (b << 5) | (b >>> 27);
+        a = t + f + a - 899497514 + blocks[j + 4] << 0;
+        c = (c << 30) | (c >>> 2);
+      }
+
+      h0 = h0 + a << 0;
+      h1 = h1 + b << 0;
+      h2 = h2 + c << 0;
+      h3 = h3 + d << 0;
+      h4 = h4 + e << 0;
+    } while(!end);
+
+    return HEX_CHARS[(h0 >> 28) & 0x0F] + HEX_CHARS[(h0 >> 24) & 0x0F] +
+           HEX_CHARS[(h0 >> 20) & 0x0F] + HEX_CHARS[(h0 >> 16) & 0x0F] +
+           HEX_CHARS[(h0 >> 12) & 0x0F] + HEX_CHARS[(h0 >> 8) & 0x0F] +
+           HEX_CHARS[(h0 >> 4) & 0x0F] + HEX_CHARS[h0 & 0x0F] +
+           HEX_CHARS[(h1 >> 28) & 0x0F] + HEX_CHARS[(h1 >> 24) & 0x0F] +
+           HEX_CHARS[(h1 >> 20) & 0x0F] + HEX_CHARS[(h1 >> 16) & 0x0F] +
+           HEX_CHARS[(h1 >> 12) & 0x0F] + HEX_CHARS[(h1 >> 8) & 0x0F] +
+           HEX_CHARS[(h1 >> 4) & 0x0F] + HEX_CHARS[h1 & 0x0F] +
+           HEX_CHARS[(h2 >> 28) & 0x0F] + HEX_CHARS[(h2 >> 24) & 0x0F] +
+           HEX_CHARS[(h2 >> 20) & 0x0F] + HEX_CHARS[(h2 >> 16) & 0x0F] +
+           HEX_CHARS[(h2 >> 12) & 0x0F] + HEX_CHARS[(h2 >> 8) & 0x0F] +
+           HEX_CHARS[(h2 >> 4) & 0x0F] + HEX_CHARS[h2 & 0x0F] +
+           HEX_CHARS[(h3 >> 28) & 0x0F] + HEX_CHARS[(h3 >> 24) & 0x0F] +
+           HEX_CHARS[(h3 >> 20) & 0x0F] + HEX_CHARS[(h3 >> 16) & 0x0F] +
+           HEX_CHARS[(h3 >> 12) & 0x0F] + HEX_CHARS[(h3 >> 8) & 0x0F] +
+           HEX_CHARS[(h3 >> 4) & 0x0F] + HEX_CHARS[h3 & 0x0F] +
+           HEX_CHARS[(h4 >> 28) & 0x0F] + HEX_CHARS[(h4 >> 24) & 0x0F] +
+           HEX_CHARS[(h4 >> 20) & 0x0F] + HEX_CHARS[(h4 >> 16) & 0x0F] +
+           HEX_CHARS[(h4 >> 12) & 0x0F] + HEX_CHARS[(h4 >> 8) & 0x0F] +
+           HEX_CHARS[(h4 >> 4) & 0x0F] + HEX_CHARS[h4 & 0x0F];
   };
 
-  var getBlocksFromAscii = function(message) {
-    // a block is 32 bits(4 bytes), a chunk is 512 bits(64 bytes)
-    var length = message.length;
-    var chunkCount = ((length + 8) >> 6) + 1;
-    var blockCount = chunkCount << 4; // chunkCount * 16
-    var blocks = [], i;
-    for(i = 0;i < blockCount;++i) {
-      blocks[i] = 0;
-    }
-    for(i = 0;i < length;++i) {
-      blocks[i >> 2] |= message.charCodeAt(i) << (3 - (i & 3) << 3);
-    }
-    blocks[i >> 2] |= 0x80 << (3 - (i & 3) << 3);
-    blocks[blockCount - 1] = length << 3; // length * 8
-    return blocks;
-  };
-
-  var getBlocksFromUtf8 = function(message) {
-    var bytes = getBytesFromUtf8(message);
-    var length = bytes.length;
-    var chunkCount = ((length + 8) >> 6) + 1;
-    var blockCount = chunkCount << 4; // chunkCount * 16
-    var blocks = [], i;
-    for(i = 0;i < blockCount;++i) {
-      blocks[i] = 0;
-    }
-    for(i = 0;i < length;++i) {
-      blocks[i >> 2] |= bytes[i] << (3 - (i & 3) << 3);
-    }
-    blocks[i >> 2] |= 0x80 << (3 - (i & 3) << 3);
-    blocks[blockCount - 1] = length << 3; // length * 8
-    return blocks;
-  };
-
-  if(typeof(module) != 'undefined') {
+  if(!root.JS_SHA1_TEST && typeof(module) != 'undefined') {
     module.exports = sha1;
   }
   else if(root) {